+++ scipy/stats/meson.build
@@ -182,88 +182,6 @@ py3.install_sources([
   subdir: 'scipy/stats'
 )
 
-_ufuncs_pxi_pxd_sources = [
-  fs.copyfile('_cython_stats.pxi'),
-  fs.copyfile('_ufuncs_extra_code.pxi'),
-  fs.copyfile('_ufuncs_extra_code_common.pxi')
-]
-
-# TODO: this installs all generated files, while we want to install only:
-#       _ufuncs.pyi, _cython_stats.pxd
-#       There's no good way around this currently, because py3.install_sources
-#       doesn't accept generated targets. See TODO near the end of
-#       scipy/linalg/meson.build for more details.
-cython_stats = custom_target('cython_stats',
-  output: [
-    '_ufuncs.pyx',
-    '_ufuncs_defs.h',
-    '_ufuncs_cxx.pyx',
-    '_ufuncs_cxx.pxd',
-    '_ufuncs_cxx_defs.h',
-    '_ufuncs.pyi',
-    'cython_stats.pyx',
-    'cython_stats.pxd'
-  ],
-  input: ['_generate_pyx_.py', 'functions.json', '_add_newdocs.py'],
-  command: [py3, '@INPUT0@', '-o', '@OUTDIR@'],
-  install: true,
-  install_dir: py3.get_install_dir() / 'scipy/stats',
-  install_tag: 'devel',
-)
-
-# Only needed to establish a dependency; see comments in linalg/meson.build
-# for cython_blas_pxd
-cython_stats_pxd = custom_target('_dummy_cython_stats.pxd',
-  output: '_dummy_cython_stats.pxd',
-  input: cython_stats[7],
-  command: [py3, '-c', '"@INPUT@"'],
-)
-
-# pyx -> c, pyx -> cpp generators, depending on copied pxi, pxd files.
-uf_cython_gen = generator(cython,
-  arguments : cython_args,
-  output : '@BASENAME@.c',
-  depends : [_cython_tree, _ufuncs_pxi_pxd_sources, cython_stats_pxd, _stats_pxd[0]])
-
-uf_cython_gen_cpp = generator(cython,
-  arguments : cython_cplus_args,
-  output : '@BASENAME@.cpp',
-  depends : [_cython_tree, _ufuncs_pxi_pxd_sources, cython_stats_pxd, _stats_pxd[0]])
-
-py3.extension_module('_ufuncs',
-  [
-    uf_cython_gen.process(cython_stats[0]),  # _ufuncs.pyx
-  ],
-  c_args: [cython_c_args, Wno_maybe_uninitialized],
-  include_directories: ['../_lib', '../_build_utils/src'],
-  dependencies: [
-    lapack,
-    npymath_lib,
-    np_dep,
-  ],
-  install: true,
-  subdir: 'scipy/stats'
-)
-
-ufuncs_cxx_cpp_args = [
-  cython_cpp_args,
-  '-DBOOST_MATH_STANDALONE=1',
-  '-DCYTHON_EXTERN_C=extern "C"',
-]
-
-py3.extension_module('_ufuncs_cxx',
-  [
-    uf_cython_gen_cpp.process(cython_stats[2]),  # _ufuncs_cxx.pyx
-    ],
-  cpp_args: ufuncs_cxx_cpp_args,
-  include_directories: ['../_lib/boost_math/include', '../_lib',
-                        '../_build_utils/src'],
-  link_args: version_link_args,
-  dependencies: [np_dep],
-  install: true,
-  subdir: 'scipy/stats'
-)
-
 subdir('_levy_stable')
 subdir('_unuran')
 subdir('_rcont')
